+++ src/js/components/layout/Default.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import AppLayout from './App';
-import AdaccountSelector from '../../containers/AdaccountSelector';
+import AdAccountSelector from '../../containers/AdAccountSelector';
 import AdSetSelector from '../../containers/AdSetSelector';
 import SubmitForm from '../../containers/SubmitForm';
 
 const Layout = (props) => {
-  const { name, email, currentAdaccountId } = props;
+  const { name, email, currentAdAccountId } = props;
   let form;
   let adSet;
-  if (currentAdaccountId) {
+  if (currentAdAccountId) {
     adSet = (
       <div style={{ marginBottom: '20px' }}>
-        <AdSetSelector currentAdAccountId={currentAdaccountId} />
+        <AdSetSelector currentAdAccountId={currentAdAccountId} />
       </div>
     );
 
@@ -27,7 +27,7 @@ const Layout = (props) => {
       <h3>Hello {name} ({email})</h3>
       <div>
         <div style={{ marginBottom: '20px' }}>
-          <AdaccountSelector />
+          <AdAccountSelector />
         </div>
         {adSet}
         {form}
